+++ packages/simcore-sdk/src/simcore_sdk/node_ports_v2/port.py
@@ -1,12 +1,12 @@
 import logging
 from pathlib import Path
 from pprint import pformat
-from typing import Dict, Optional, Tuple, Type, Union
+from typing import Dict, Optional, Tuple, Type
 
 from models_library.services import PROPERTY_KEY_RE, ServiceProperty
-from pydantic import BaseModel, Field, PrivateAttr, validator
+from pydantic import Field, PrivateAttr, validator
 
-from ..node_ports.exceptions import InvalidItemTypeError, UnboundPortError
+from ..node_ports.exceptions import InvalidItemTypeError
 from . import port_utils
 from .links import DataItemValue, DownloadLink, FileLink, ItemConcreteValue, PortLink
 
@@ -114,35 +114,3 @@ class Port(ServiceProperty):
 
         self.value = converted_value
         await self._node_ports.save_to_db_cb(self._node_ports)
-
-
-PortKey = str
-
-
-class PortsMapping(BaseModel):
-    __root__: Dict[PortKey, Port]
-
-    def __getitem__(self, key: Union[int, PortKey]) -> Port:
-        if isinstance(key, int):
-            if key < len(self.__root__):
-                key = list(self.__root__.keys())[key]
-        if not key in self.__root__:
-            raise UnboundPortError(key)
-        return self.__root__[key]
-
-    def __iter__(self):
-        return iter(self.__root__)
-
-    def items(self):
-        return self.__root__.items()
-
-    def values(self):
-        return self.__root__.values()
-
-
-class InputsList(PortsMapping):
-    __root__: Dict[PortKey, Port]
-
-
-class OutputsList(PortsMapping):
-    __root__: Dict[PortKey, Port]
